+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Adjusted filter function to stringify decoded message data of Long types for correct comparison with filters. (#173)
 
 ## [2.10.2] - 2023-08-24
 ### Changed